requirements.txt
huggingface_hub>=0.20.0
numpy>=1.21.0
python-dotenv>=1.0.0
transformers>=4.35.0
fastapi>=0.110.0
uvicorn[standard]>=0.22.0
reportlab>=4.0.0
python-multipart>=0.0.6
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper>=20231117
SpeechRecognition>=3.10.0
pydub>=0.25.1
langchain>=0.1.16
langchain-community>=0.0.21
langchain-text-splitters>=0.0.1
pypdf2>=3.0.1
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
chromadb>=0.4.22
Sastrawi>=1.0.1
tqdm>=4.66.0
scikit-learn>=1.3.0
pandas>=2.0.0
gensim>=4.3.0
nltk>=3.8.1
beautifulsoup4>=4.12.2
selenium>=4.10.0
webdriver-manager>=4.0.0
librosa>=0.10.0
ipywidgets>=8.0.0
evaluate>=0.4.0
jiwer>=3.0.3
fasttext-wheel>=0.9.2
openpyxl>=3.1.2
scipy>=1.10.0
gradio>=4.16.0
tensorboard>=2.15.1
datasets>=2.18.0
Pillow>=10.0.0
pytesseract>=0.3.10